+++ NEWS.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * `Distributions.jl`, `RecursiveArrayTools.jl` and `HybridArrays.jl` were moved to weak dependencies to reduce load time and improve extensibility.
-* `translate_diff`, `inv_diff` and thus `apply_diff_group`, are available for all the groups with invariant tangent vector storage
+* `translate_diff`, `inv_diff` and thus `apply_diff_group`, are available for all the groups with invariant tangent vector storage.
+* `SpecialEuclidean` group now has a different default tangent vector representation, the left-invariant one; to get the old representation pass `gvr=HybridTangentRepresentation()` to the constructor of `SpecialEuclidean`.
 * `adjoint_action` takes a direction argument; by default it is `LeftAction`.
 * `adjoint_action!` is the necessary method to implement in groups with left-invariant tangent vector representation.
 * On groups with invariant vector storage, `exp` and `log` are based on `exp_lie` and `log_lie` and thus group invariant.